Eduardo moreno

+++ src/main/java/com/vanguardiapropiedades/inmobiliaria/servicios/UsuarioServicio.java
@@ -22,6 +22,7 @@ import org.springframework.web.context.request.ServletRequestAttributes;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.vanguardiapropiedades.inmobiliaria.entidades.ImagenEntidad;
+import com.vanguardiapropiedades.inmobiliaria.entidades.PropiedadEntidad;
 import com.vanguardiapropiedades.inmobiliaria.entidades.UsuarioEntidad;
 import com.vanguardiapropiedades.inmobiliaria.Enums.Rol;
 import com.vanguardiapropiedades.inmobiliaria.excepciones.MiException;
@@ -43,7 +44,7 @@ public class UsuarioServicio implements UserDetailsService {
     public void crearUsuario(String nombre, String dni, String email, String password, String password2)
             throws MiException {
         UsuarioEntidad user = new UsuarioEntidad();
-        // validar(nombre, email, password, password2);
+        validar(nombre, dni, email, password, password2);
         user.setNombre(nombre);
         user.setEmail(email);
         user.setDni(dni);
@@ -63,12 +64,13 @@ public class UsuarioServicio implements UserDetailsService {
 
     // UPDATE
     public void editarUsuario(String id, String dni, String nombre, String email, String password, String password2,
-            MultipartFile foto)
+            MultipartFile foto, String rol)
             throws MiException {
         Optional<UsuarioEntidad> respuesta = usuarioRepositorio.findById(id);
         if (respuesta.isPresent()) {
             UsuarioEntidad user = respuesta.get();
-            validar(nombre, email, password, password2);
+            // ? Validar datos
+            // validar(nombre, dni, email, password, password2);
             // ? Verificar si el usuario tiene foto
             if (user.getImagen() == null) {
                 if (foto.getSize() > 0) {
@@ -85,6 +87,38 @@ public class UsuarioServicio implements UserDetailsService {
             user.setEmail(email);
             user.setDni(dni);
             user.setPassword(new BCryptPasswordEncoder().encode(password));
+            user.setRol(Rol.valueOf(rol));
+            usuarioRepositorio.save(user);
+        }
+    }
+
+    // UPDATE
+    public void editarUsuarioAdmin(String id, String dni, String nombre, String email, String password,
+            String password2,
+            MultipartFile foto, String rol)
+            throws MiException {
+        Optional<UsuarioEntidad> respuesta = usuarioRepositorio.findById(id);
+        if (respuesta.isPresent()) {
+            UsuarioEntidad user = respuesta.get();
+            // ? Validar datos
+            // validar(nombre, dni, email, password, password2);
+            // ? Verificar si el usuario tiene foto
+            if (user.getImagen() == null) {
+                if (foto.getSize() > 0) {
+                    ImagenEntidad img = imagenServicio.crearImagen(foto);
+                    user.setImagen(img);
+                }
+            } else {
+                if (foto.getSize() > 0) {
+                    ImagenEntidad img = imagenServicio.editarImagen(foto, user.getImagen().getId());
+                    user.setImagen(img);
+                }
+            }
+            user.setNombre(nombre);
+            user.setEmail(email);
+            user.setDni(dni);
+            user.setPassword(password);
+            user.setRol(Rol.valueOf(rol));
             usuarioRepositorio.save(user);
         }
     }
@@ -92,23 +126,35 @@ public class UsuarioServicio implements UserDetailsService {
     // DELETE
     @Transactional
     public void eliminarUsuario(String id) throws MiException {
-        usuarioRepositorio.deleteById(id);
+        UsuarioEntidad usuario = buscarPorId(id).get();
+        if (usuario.getPropiedades().isEmpty()) {
+            usuarioRepositorio.deleteById(id);
+        }
     }
 
-    // TODO: Agregar DNI
-    private void validar(String nombre, String email, String password, String password2) throws MiException {
-
-        // verificar que el email sea valido
-        String regex = "([a-z0-9]+(\\.?[a-z0-9])*)+@(([a-z0-9]+)\\.([a-z0-9]+))+"; // expresion regular
-        Pattern pattern = Pattern.compile(regex); // compilar la expresion regular
+    private void validar(String nombre, String dni, String email, String password, String password2)
+            throws MiException {
 
         if (nombre.isEmpty() || nombre.isBlank()) {
             throw new MiException("El nombre no puede estar vacio");
         }
+        // Verifica si dni está vacío y sea numérico
+        if (dni.isEmpty()) {
+            throw new MiException("El dni no puede estar vacio");
+        }
+        // Verifica si dni es numérico parseando el String dni a Integer
+        try {
+            Integer.parseInt(dni);
+        } catch (NumberFormatException e) {
+            throw new MiException("El dni debe contener solo numéros");
+        }
+        // Verifica si email está vacío
         if (email.isEmpty()) {
             throw new MiException("El email no puede estar vacio");
         }
-
+        // Verifica email formato correo
+        String regex = "([a-z0-9]+(\\.?[a-z0-9])*)+@(([a-z0-9]+)\\.([a-z0-9]+))+"; // expresion regular
+        Pattern pattern = Pattern.compile(regex); // compilar la expresion regular
         if (!pattern.matcher(email).matches()) {
             throw new MiException("El email no es valido");
         }
@@ -116,10 +162,14 @@ public class UsuarioServicio implements UserDetailsService {
         if (password.isEmpty()) {
             throw new MiException("La contraseña no puede estar vacia");
         }
+        // Verifica si la contraseña tiene mas de 6 caracteres
+        if (password.length() < 6) {
+            throw new MiException("La contraseña debe tener mas de 6 caracteres");
+        }
+        // Verifica si las contraseñas coinciden
         if (!password.equals(password2)) {
             throw new MiException("La contraseñas no coinciden");
         }
-
     }
 
     @Override
@@ -154,5 +204,28 @@ public class UsuarioServicio implements UserDetailsService {
 
     public Page<UsuarioEntidad> listarUsuarios(Pageable pageable) {
         return usuarioRepositorio.findAll(pageable);
+    }
+
+    public List<PropiedadEntidad> propiedadesUsuario(String id) {
+        UsuarioEntidad user = buscarPorId(id).get();
+        return user.getPropiedades();
+    }
+
+    // ? Metodos buscar usuario para el administrador
+    public Page<UsuarioEntidad> listarUsuariosAdmin(String dni, Pageable pageable) {
+        return usuarioRepositorio.findByDni(dni, pageable);
+    }
+
+    public Page<UsuarioEntidad> buscarPorEmail(String email, Pageable pageable) {
+        return usuarioRepositorio.buscarporEmail(email, pageable);
+    }
+    public void cambiarRol(String id){
+        UsuarioEntidad user = buscarPorId(id).get();
+        if (user.getRol().toString().equals("CLIENT")) {
+            user.setRol(Rol.ENTE);
+        }else{
+            user.setRol(Rol.CLIENT);
+        }
+        usuarioRepositorio.save(user);
     }
 }
